requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.1.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.4
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==2.10
itsdangerous==2.0.1
kombu==5.1.0
numpy==1.20.3
pandas==1.2.4
prompt-toolkit==3.0.18
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-cache==0.6.4
six==1.16.0
sqlparse==0.4.1
url-normalize==1.4.3
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
